requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
asyncpg
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
apscheduler
email-validator
python-dotenv
